How to Structure Directories/Files(OPTIONAL)
OPTIONAL
Chapter 1 is how the rest of the chapters should be structured if you want to follow this route. So you have One repo that contains all the classwork and lecture/chapter code. Much easier to manage on your end and much easier to grade on mine, thank you in advance!