OPTIONAL
OPTIONAL but looks good to have a "Clean" github, I even suggest making one repo called "MCTC" and inside have a folder for each class where you write code. A repo is just a folder that git is tracking on your local computer and GitHub is the "cloud" where you upload(push) your files, folders(directories), and projects(in our case is a directory that contains all the files for a C# .NET Core Console Application.